+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Unreleased
 
 #### Changed
 
+-   Switched to ISO 639-3 language codes. (\#468)
 -   Updated scraped data in preparation for the SIGMORPHON 2022 shared task:
     `swe nno ger dut ita rum ukr bel tgl ceb ben asm per pus tha lwl`. (\#461)
 -   Made scripts under `data/frequencies/` and `data/morphology/` more flexible,
@@ -108,6 +109,10 @@ Unreleased
 #### Removed
 
 -  Dropped support for Python 3.6. (\#462)
+
+#### Changed
+
+-   Switched to ISO 639-3 language codes. (\#468)
 
 [1.2.0] - 2021-01-30
 --------------------